\name{VBspPCA-package}
\alias{VBspPCA-package}
\alias{VBspPCA}
\docType{package}
\title{
\packageTitle{VBspPCA}
}
\description{
\packageDescription{VBspPCA}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{VBspPCA}

Maintainer: \packageMaintainer{VBspPCA}
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = ssnmf)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{ssnmf}}
%\section{Documentation Index}{\packageIndices{ssnmf}}

\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
